fix: remove duplicated file for cache unit tests
